Moved all `index.js` code to the `src/cachimo.js`
