+++ tests/promise.test.js
@@ -1,4 +1,4 @@
-const cachimo = require('../index');
+const cachimo = require('../src/cachimo');
 
 test('promise', async () => {
   try {
